requirements.txt
black>=20.8b1
click>=7.0
loguru>=0.5.3
Markdown>=3.1.1
numpy>=1.16.3
opencv-python>=4.4.0.46
pandas>=0.24.2
pathlib>=1.0.1
Pillow>=6.0.0
scikit-image>=0.17.2
scikit-learn>=0.23.2
tensorboard>=2.4.0
torch>=1.1.0
torchvision>=0.3.0
tqdm>=4.32.2
wand>=0.6.5
learn2learn>=0.1.0
ray[tune]
seaborn